% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_fivethirtyeight.R
\name{theme_538}
\alias{theme_538}
\title{538 ggplot theme}
\usage{
theme_538(base_size = 13, base_family = "")
}
\arguments{
\item{base_size}{size of font}

\item{base_family}{font used}
}
\description{
538 ggplot theme
}
\details{
a 538 theme for ggplot
}
\examples{
ggplot(mtcars, aes(x = mpg)) + geom_histogram() + theme_538()
}
\concept{themes theme_538}
