% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dark_blue.R
\name{theme_dark_blue}
\alias{theme_dark_blue}
\title{Title}
\usage{
theme_dark_blue(base_size = 14, base_family = "sans")
}
\arguments{
\item{base_size}{size of font}

\item{base_family}{what font}
}
\description{
Title
}
