% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/white_theme.R
\name{white_theme}
\alias{white_theme}
\title{Theme inspired by ggthemes::theme_fivethirtyeight
fivethirtyeight.com}
\usage{
white_theme(base_size = 12, font = "Lato")
}
\arguments{
\item{base_size}{size of font}

\item{font}{font you want to use}
}
\description{
Theme inspired by ggthemes::theme_fivethirtyeight
fivethirtyeight.com
}
